// src/store/settings/types.ts
export enum AppTabs {
  publications = "Publications",
  trackings = "Trackings",
  users = "Users",
}

export enum SortModes {
  // Публикации
  addDate = "Дата добавления",
  pubName = "Имя публикации",
  pubDate = "Дата публикации",

  // Трекинги
  dateIn = "Дата выдачи",
  dateOut = "Дата сдачи",
  takerName = "Держатель",

  // Пользователи
  fullName = "Ф. И. О.",
  department = "Отдел",
  role = "Роль",
}